refactor: matching general build plans page with figma design
