+++ src/pages/saved-project-plans.jsx
@@ -1,33 +1,48 @@
-import { Box, Button, Flex, Heading, VStack } from "@chakra-ui/react";
-import Link from "next/link";
-import { useEffect, useState } from "react";
-import { getPlans } from "src/actions/Plan";
+import { Box, Button, Flex, Heading, Spacer, VStack } from "@chakra-ui/react";
 import SavedProjectPlan from "src/components/SavedProjectPlan/SavedProjectPlan";
 
 const SavedProjectPlans = () => {
-  const [plans, setPlans] = useState([]);
+  // const [plans, setPlans] = useState([]);
 
-  useEffect(() => {
-    getPlans().then((res) => {
-      setPlans(res);
-    });
-  }, []);
+  // useEffect(() => {
+  //   getPlans().then((res) => {
+  //     setPlans(res);
+  //   });
+  // }, []);
 
   return (
     <Box>
       <Flex mb="10">
-        {/* This needs to use the NavLink component */}
-        <Link href="/plan-builder">
-          <Button>Return to Project Plan Builder</Button>
-        </Link>
-        <Heading m="auto" ml="25%">
-          Saved Project Plans
-        </Heading>
+        <Button
+          as="a"
+          href="/plan-builder"
+          background="#6D6E70"
+          borderRadius="3xl"
+          color="white"
+          marginBottom="20"
+        >
+          Return to Project Plan Builder Home
+        </Button>
+        <Spacer />
+        <Heading>Saved Project Plans</Heading>
+        <Spacer />
+        <Button
+          as="a"
+          href="/plan-builder"
+          background="#6D6E70"
+          borderRadius="3xl"
+          color="white"
+          marginBottom="20"
+          visibility="hidden"
+        >
+          Return to Project Plan Builder Home
+        </Button>
       </Flex>
       <VStack p="0% 2% 0% 2%">
-        {plans.map((plan, index) => {
+        {/* {plans.map((plan, index) => {
           return <SavedProjectPlan key={index} plan={plan} />;
-        })}
+        })} */}
+        return <SavedProjectPlan />
       </VStack>
     </Box>
   );
